requirements.txt
APScheduler==3.10.4
asgiref==3.5.2
beautifulsoup4==4.12.3
cached-property==1.5.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.1.1
django-adminlte-3==0.1.6
django-adminlte3==0.1.6
django-apscheduler==0.6.2
django-bootstrap-form==3.4
django-cors-headers==3.13.0
django-crispy-forms==2.0
django-crontab==0.7.1
django-dbbackup==4.0.2
django-dotenv==1.4.2
django-filter==21.1
django-import-export==3.3.7
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
django-url-filter==0.3.15
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
engineering-notation==0.6.0
enum-compat==0.0.3
et-xmlfile==1.1.0
ffmpeg-python==0.2.0
flatbuffers==22.10.26
fpdf==1.7.2
future==1.0.0
gunicorn==20.1.0
idna==3.6
install==1.3.5
lxml==5.2.2
MarkupPy==1.14
numpy==1.26.4
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.2.2
phonenumbers==8.13.33
pillow==10.3.0
pycparser==2.22
PyJWT==2.6.0
python-dateutil==2.9.0.post0
pytz==2022.4
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.3
tablib==3.5.0
tzdata==2022.4
tzlocal==5.2
urllib3==2.2.1
xlrd==2.0.1
XlsxWriter==1.1.8
xlwt==1.3.0
zipp==3.10.0
